src/main/java/com/itexps/junit/project/source/Movie.java
/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
 */
package com.itexps.junit.project.source;

/**
 *
 * @author adrit
 */
public class Movie {
    
    private int id, year;
    private String name;

    public Movie(int id, int year, String name) {
        this.id = id;
        this.year = year;
        this.name = name;
    }

    public int getId() {
        return id;
    }

    public int getYear() {
        return year;
    }

    public String getName() {
        return name;
    }
    
    
}
